src/master/requirements.txt
python-telegram-bot[rate-limiter]==21.1.1
PyYAML==6.0
# pymongo==4.6.2
python-dotenv==0.21.0
# charade==1.0.3
motor==3.3.2
python-telegram-broadcast==0.9.7
